backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pandas
geopandas
shapely
python-dotenv
